requirements.txt
streamlit
pdfplumber
python-docx
google-generativeai
python-dotenv
requests
pandas
beautifulsoup4